+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For details about compatibility between different releases, see the **Commitment
 ### Added
 
 - New `ListBands` RPC on the `Configuration` service.
+- Support for loading end device template from Device Repository when importing devices using a CSV file.
 
 ### Changed